Updates app user flow
Updates app to have user start on main screen and then allows them to create and account or log in.
