+++ html/forgot-password.html
@@ -24,7 +24,7 @@
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.css"
     />
-    <link rel="stylesheet" href="../styles/app.css" />
+    <link rel="stylesheet" href="../styles/login.css" />
     <link rel="stylesheet" href="../styles/index.css" />
     <!-- Scripts  -->
     <script src="../scripts/forgot-password.js" type="module" defer></script>
@@ -32,7 +32,9 @@
 
   <body>
     <header>
-      <h1 class="title">Twimba</h1>
+      <a class="title-link" href="../index.html">
+        <h1 class="title">Twimba</h1>
+      </a>
     </header>
 
     <main>
@@ -49,7 +51,7 @@
             Send Reset Email
           </button>
         </form>
-        <a href="../index.html">Log in</a>
+        <a href="login.html">Log in</a>
         <a href="create-account.html">Create an account</a>
       </section>
     </main>
