+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Twimba | Log In</title>
+    <title>Twimba</title>
     <!-- Font Awesome  -->
     <link
       rel="stylesheet"
@@ -24,37 +24,41 @@
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.css"
     />
-    <link rel="stylesheet" href="styles/app.css" />
     <link rel="stylesheet" href="styles/index.css" />
     <!-- Scripts  -->
     <script src="scripts/index.js" type="module" defer></script>
   </head>
 
   <body>
-    <header>
-      <h1 class="title">Twimba</h1>
+    <header class="flex header">
+      <a class="title-link" href="../index.html">
+        <h1 class="title">Twimba</h1>
+      </a>
+      <button id="sign-out-btn" class="btn btn--secondary hidden">
+        Sign Out
+      </button>
     </header>
 
     <main>
-      <section id="login">
-        <h2>Log in</h2>
-        <form id="loginForm">
-          <ul>
-            <li>
-              <label for="loginEmail">Email</label>
-              <input type="email" id="loginEmail" required />
-            </li>
-            <li>
-              <label for="loginPassword">Password</label>
-              <input type="password" id="loginPassword" required />
-            </li>
-            <li>
-              <a href="html/forgot-password.html">Forgot password</a>
-            </li>
-          </ul>
-          <button class="btn btn--primary" type="submit">Login</button>
-        </form>
-        <a href="html/create-account.html">Create an account</a>
+      <h3 class="status-message" id="status-message"></h3>
+      <section id="tweet-input">
+        <div class="tweet-input__container flex">
+          <img
+            src="images/scrimbalogo.png"
+            class="profile-pic"
+            alt="Scrimba profile picture"
+          />
+          <textarea
+            id="tweet-textarea"
+            class="tweet-textarea"
+            placeholder="What's happening?"
+          ></textarea>
+        </div>
+        <button id="tweet-btn" class="btn btn--primary">Tweet</button>
+      </section>
+
+      <section id="tweet-feed" class="tweet-feed">
+        <!-- Where Firebase data gets rendered -->
       </section>
     </main>
   </body>
